cleaned up cryptography lib and tests
